+++ src/assets/UpdatedLogs.md
@@ -1,0 +1,47 @@
+### v0.3.6 (2023-01-08)
+
+优化：剪切和清空输入文本时不再清空输出框 ([评论区建议](https://www.bilibili.com/video/BV1T24y1X78w/#reply145757356944))
+
+### v0.3.5 (2023-01-07)
+
+优化：文本框下新增清空按钮 ([评论区建议](https://www.bilibili.com/video/BV1T24y1X78w/#reply145609989120))
+
+### v0.3.4 (2023-01-06)
+
+新增功能：删除小数点、冒号与数字之间的空格 ([评论区建议](https://www.bilibili.com/video/BV1T24y1X78w/#reply145368905056))
+
+新增：更新日志展示页面
+
+优化：对输入文本的监测方式
+
+### v0.3.3 (2022-12-27)
+
+修复：部分替换只能激活一次
+
+### v0.3.2 (2022-12-22)
+
+新增功能：中英文标点互转
+
+新增功能：在字母与数字之间添加空格
+
+### v0.3.1 (2022-12-21)
+
+优化：在偏好设置处显示"启用的功能数量/总数"
+
+修复：复制成功的提示未能正确显示
+
+### v0.3.0 (2022-12-12)
+
+新增功能：可缓存功能的激活状态
+
+新增功能：可调整功能激活顺序
+
+### v0.2 (2022-12-07)
+
+新增：可在偏好设置中按需启用功能
+
+新增：在输入框中内置 Grammarly 语法检测功能
+
+### v0.1 (2022-08-12)
+
+新增功能：可去除空格、换行符
